gzReader
dont use html tmpl

update readme; remove html tmpl

update readme

代码优化;解决添加字典时候,默认字典没变更问题

toSlash

代码优化

web handler

LocalCache.defaultDictBasePath

padding

update readme
