+++ mywords-flutter/lib/pages/lookup_word.dart
@@ -6,6 +6,8 @@ import 'package:mywords/libso/handler.dart';
 import 'package:mywords/libso/resp_data.dart';
 import 'package:mywords/widgets/word_common.dart';
 
+import '../util/local_cache.dart';
+
 class LoopUpWord extends StatefulWidget {
   const LoopUpWord({super.key});
 
@@ -46,7 +48,9 @@ class _State extends State<LoopUpWord> with AutomaticKeepAliveClientMixin {
       return;
     }
     final RespData<List<String>> respData;
-    final defaultDict = (await handler.getDefaultDict()).data ?? '';
+    LocalCache.defaultDictBasePath ??=
+        ((await handler.getDefaultDict()).data ?? '');
+    final defaultDict = LocalCache.defaultDictBasePath ?? '';
     if (defaultDict == "") {
       respData = await handler.searchByKeyWordWithDefault(v);
     } else {
