+++ mywords-flutter/lib/widgets/line_chart.dart
@@ -1,7 +1,6 @@
 import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:mywords/common/prefs/prefs.dart';
-import 'package:mywords/util/util.dart';
 
 class LineValue {
   late final String tip; //图例提示文字
