--- mywords-go/cmd/flutter.go
@@ -1,1 +1,0 @@
-package main
